move utility methods out of API package
