--- logstash-core/src/test/java/co/elastic/logstash/api/PluginHelperTest.java
@@ -1,11 +1,0 @@
-package co.elastic.logstash.api;
-
-import org.junit.Test;
-
-public class PluginHelperTest {
-
-    @Test
-    public void testPluginId() {
-    }
-
-}
